+++ src/Exception/CommandAbortedException.php
@@ -1,0 +1,7 @@
+<?php declare(strict_types=1);
+
+namespace Becklyn\DeployMessageGenerator\Exception;
+
+class CommandAbortedException extends \RuntimeException implements DeployMessageGeneratorExceptionInterface
+{
+}