+++ CHANGELOG.md
@@ -1,8 +1,9 @@
-vNext
+2.2.0
 =====
 
 *   (internal) Replace CircleCI with GitHub Actions.
 *   (internal) Add proper support for running in PHP 7.4 with Symfony 5 and PHP >8 with Symfony 6.
+*   (feature) Added new Command `ticket:show-ticket-information`, which will display the ticket information from the given Commit Range.
 
 
 2.1.0